New crowns for 2015 and 2019; DEM trend removal
The 2015 data had a much more pronounced slope across the garden. Attempted to model that trend (DEM ~ x + y) and then automate the crown segmentation with the residuals. Might need to fine tune the thresholds since this generated more small polygons and split the larger ones. Code included for generating the segmented crowns.